# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Etsy Mockup Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export interface Product {
  id: string;
  name: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateProductBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface ProductStats {
  sessionCount: number;
  templateCount: number;
  completedSessionCount: number;
  /** @nullable */
  lastSessionAt: string | null;
}

export type SessionFalParams = { [key: string]: unknown };

export interface QAAnswer {
  question: string;
  answer: string;
  questionIndex: number;
}

export interface Session {
  id: string;
  productId: string;
  /**
   * A or B
   * @nullable
   */
  optionType: string | null;
  /**
   * M1 or M2
   * @nullable
   */
  outputType: string | null;
  /**
   * For M2: how many images (2-8)
   * @nullable
   */
  imageCount: number | null;
  /**
   * SAME or IDEA (Option B only)
   * @nullable
   */
  referenceStyle: string | null;
  /**
   * 1-100 for SAME mode
   * @nullable
   */
  similarityLevel: number | null;
  productImageUrls: string[];
  /** @nullable */
  referenceImageUrl: string | null;
  /** @nullable */
  referenceAnalysis: string | null;
  qaAnswers: QAAnswer[];
  /** @nullable */
  finalPrompt: string | null;
  /** @nullable */
  enhancedPrompt: string | null;
  /** @nullable */
  falModelId: string | null;
  falParams: SessionFalParams;
  generatedImageUrls: string[];
  /** draft | prompt_ready | generating | completed | failed */
  status: string;
  /** @nullable */
  templateInspirationId: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSessionBody {
  productId: string;
  optionType: string;
  outputType: string;
  /** @nullable */
  imageCount?: number | null;
  /** @nullable */
  referenceStyle?: string | null;
  /** @nullable */
  similarityLevel?: number | null;
  productImageUrls: string[];
  /** @nullable */
  referenceImageUrl?: string | null;
  /** @nullable */
  templateInspirationId?: string | null;
}

export type UpdateSessionBodyFalParams = { [key: string]: unknown };

export interface UpdateSessionBody {
  /** @nullable */
  optionType?: string | null;
  /** @nullable */
  outputType?: string | null;
  /** @nullable */
  imageCount?: number | null;
  /** @nullable */
  referenceStyle?: string | null;
  /** @nullable */
  similarityLevel?: number | null;
  productImageUrls?: string[];
  /** @nullable */
  referenceImageUrl?: string | null;
  qaAnswers?: QAAnswer[];
  /** @nullable */
  finalPrompt?: string | null;
  /** @nullable */
  enhancedPrompt?: string | null;
  /** @nullable */
  falModelId?: string | null;
  falParams?: UpdateSessionBodyFalParams;
  generatedImageUrls?: string[];
  /** @nullable */
  status?: string | null;
  /** @nullable */
  templateInspirationId?: string | null;
}

export interface ReferenceAnalysis {
  analysis: string;
  sessionId: string;
}

export interface QAOption {
  label: string;
  description: string;
}

export interface QAQuestion {
  done: boolean;
  /** @nullable */
  question?: string | null;
  options?: QAOption[];
  /** @nullable */
  aiSuggestion?: string | null;
  /** @nullable */
  questionIndex?: number | null;
  /**
   * Set when done=true, the AI-constructed final prompt
   * @nullable
   */
  finalPrompt?: string | null;
}

export interface SubmitAnswerBody {
  question: string;
  answer: string;
  questionIndex: number;
}

export interface EnhanceSuggestion {
  original: string;
  replacement: string;
  reason: string;
}

export interface EnhanceSuggestions {
  suggestions: EnhanceSuggestion[];
}

export interface RevisePromptBody {
  instruction: string;
}

export interface RevisedPrompt {
  prompt: string;
}

export type GenerateImagesBodyFalParams = { [key: string]: unknown };

export interface GenerateImagesBody {
  falModelId: string;
  falParams?: GenerateImagesBodyFalParams;
  /** @nullable */
  imageCount?: number | null;
}

export interface GenerationResult {
  imageUrls: string[];
  sessionId: string;
}

export type TemplateSessionConfig = { [key: string]: unknown };

export interface Template {
  id: string;
  productId: string;
  name: string;
  /** M1 or M2 */
  type: string;
  /** A or B */
  optionType: string;
  prompt: string;
  imageUrls: string[];
  sessionConfig: TemplateSessionConfig;
  createdAt: string;
}

export interface CreateTemplateBody {
  sessionId: string;
  name: string;
}

export interface Settings {
  id: string;
  falApiKeySet: boolean;
  openrouterApiKeySet: boolean;
  claudeApiKeySet: boolean;
  claudeEnabled: boolean;
}

export interface UpdateSettingsBody {
  /** @nullable */
  falApiKey?: string | null;
  /** @nullable */
  openrouterApiKey?: string | null;
  /** @nullable */
  claudeApiKey?: string | null;
  /** @nullable */
  claudeEnabled?: boolean | null;
}

export type FalModelParamsSchema = { [key: string]: unknown };

export type FalModelDefaultValues = { [key: string]: unknown };

export interface FalModel {
  id: string;
  name: string;
  endpoint: string;
  curlCommand: string;
  paramsSchema: FalModelParamsSchema;
  defaultValues: FalModelDefaultValues;
  createdAt: string;
}

export interface CreateFalModelBody {
  name: string;
  curlCommand: string;
}

export type UpdateFalModelBodyDefaultValues = { [key: string]: unknown };

export interface UpdateFalModelBody {
  name?: string;
  curlCommand?: string;
  defaultValues?: UpdateFalModelBodyDefaultValues;
}

export type LlmConfigParamsSchema = { [key: string]: unknown };

export type LlmConfigDefaultValues = { [key: string]: unknown };

export interface LlmConfig {
  id: string;
  name: string;
  /** openrouter or claude */
  provider: string;
  modelId: string;
  endpoint: string;
  /** @nullable */
  systemPrompt: string | null;
  curlCommand: string;
  paramsSchema: LlmConfigParamsSchema;
  defaultValues: LlmConfigDefaultValues;
  isActive: boolean;
  createdAt: string;
}

export interface CreateLlmConfigBody {
  name: string;
  curlCommand: string;
  /** @nullable */
  systemPrompt?: string | null;
}

export type UpdateLlmConfigBodyDefaultValues = { [key: string]: unknown };

export interface UpdateLlmConfigBody {
  name?: string;
  curlCommand?: string;
  /** @nullable */
  systemPrompt?: string | null;
  defaultValues?: UpdateLlmConfigBodyDefaultValues;
  isActive?: boolean;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListSessionsParams = {
  productId?: string;
};

export type ListTemplatesParams = {
  productId?: string;
};