# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Etsy Mockup Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sessions/{id}/qa/next-question:
    post:
      operationId: getNextQuestion
      tags: [ai]
      summary: Get the next Q&A question from the AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Next question or done signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QAQuestion"
  /sessions/{id}/qa/answer:
    post:
      operationId: submitAnswer
      tags: [ai]
      summary: Submit an answer to the current Q&A question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /sessions/{id}/prompt/enhance:
    post:
      operationId: enhancePrompt
      tags: [ai]
      summary: Enhance the current prompt (find and replace vague words)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enhancement suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnhanceSuggestions"
  /sessions/{id}/prompt/revise:
    post:
      operationId: revisePrompt
      tags: [ai]
      summary: Revise the prompt based on user instruction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisePromptBody"
      responses:
        "200":
          description: Revised prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisedPrompt"
  /sessions/{id}/prompt/rewrite:
    post:
      operationId: rewritePrompt
      tags: [ai]
      summary: Completely rewrite the prompt from scratch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rewritten prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisedPrompt"
  /sessions/{id}/generate:
    post:
      operationId: generateImages
      tags: [generation]
      summary: Start image generation for a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateImagesBody"
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated
...
    Session:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        optionType:
          type: ["string", "null"]
          description: "A or B"
        outputType:
          type: ["string", "null"]
          description: "M1 or M2"
        imageCount:
          type: ["integer", "null"]
          description: "For M2: how many images (2-8)"
        referenceStyle:
          type: ["string", "null"]
          description: "SAME or IDEA (Option B only)"
        similarityLevel:
          type: ["integer", "null"]
          description: "1-100 for SAME mode"
        productImageUrls:
          type: array
          items:
            type: string
        referenceImageUrl:
          type: ["string", "null"]
        referenceAnalysis:
          type: ["string", "null"]
        qaAnswers:
          type: array
          items:
            $ref: "#/components/schemas/QAAnswer"
        finalPrompt:
          type: ["string", "null"]
        enhancedPrompt:
          type: ["string", "null"]
        falModelId:
          type: ["string", "null"]
        falParams:
          type: object
          additionalProperties: true
        generatedImageUrls:
          type: array
          items:
            type: string
        status:
          type: string
          description: "draft | prompt_ready | generating | completed | failed"
        templateInspirationId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - productId
        - optionType
        - outputType
        - imageCount
        - referenceStyle
        - similarityLevel
        - productImageUrls
        - referenceImageUrl
        - referenceAnalysis
        - qaAnswers
        - finalPrompt
        - enhancedPrompt
        - falModelId
        - falParams
        - generatedImageUrls
        - status
        - templateInspirationId
        - createdAt
        - updatedAt
    QAAnswer:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        questionIndex:
          type: integer
      required:
        - question
        - answer
        - questionIndex
...
    QAQuestion:
      type: object
      properties:
        done:
          type: boolean
        question:
          type: ["string", "null"]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QAOption"
        aiSuggestion:
          type: ["string", "null"]
        questionIndex:
          type: ["integer", "null"]
        finalPrompt:
          type: ["string", "null"]
          description: "Set when done=true, the AI-constructed final prompt"
      required:
        - done
    QAOption:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
      required:
        - label
        - description
    SubmitAnswerBody:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        questionIndex:
          type: integer
      required:
        - question
        - answer
        - questionIndex
    EnhanceSuggestions:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/EnhanceSuggestion"
      required:
        - suggestions
    EnhanceSuggestion:
      type: object
      properties:
        original:
          type: string
        replacement:
          type: string
        reason:
          type: string
      required:
        - original
        - replacement
        - reason
    RevisePromptBody:
      type: object
      properties:
        instruction:
          type: string
      required:
        - instruction
    RevisedPrompt:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GenerateImagesBody:
      type: object
      properties:
        falModelId:
          type: string
        falParams:
          type: object
          additionalProperties: true
        imageCount:
          type: ["integer", "null"]
      required:
        - falModelId
    GenerationResult:
      type: object
      properties:
        imageUrls:
          type: array
          items:
            type: string
        sessionId:
          type: string
      required:
        - imageUrls
        - sessionId
...