# lib/api-zod/src/generated/types/enhanceSuggestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Etsy Mockup Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { EnhanceSuggestion } from "./enhanceSuggestion";

export interface EnhanceSuggestions {
  suggestions: EnhanceSuggestion[];
}