# lib/api-zod/src/generated/types/submitAnswerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Etsy Mockup Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitAnswerBody {
  question: string;
  answer: string;
  questionIndex: number;
}